+++ app/src/main/java/com/omelan/cofi/pages/settings/BackupRestore.kt
@@ -37,10 +37,11 @@ import com.omelan.cofi.utils.getDefaultPadding
 import kotlinx.coroutines.launch
 import org.json.JSONArray
 import java.nio.charset.StandardCharsets
-import java.util.*
+import java.util.Calendar
+import java.util.Locale
 
 @Composable
-fun BackupRestoreSettings(goBack: () -> Unit) {
+fun BackupRestoreSettings(goBack: () -> Unit, goToRoot: () -> Unit) {
     val context = LocalContext.current
     val snackbarState = SnackbarHostState()
     var showDefaultRecipeDialog by remember { mutableStateOf(false) }
@@ -70,7 +71,16 @@ fun BackupRestoreSettings(goBack: () -> Unit) {
                 hostState = snackbarState,
                 modifier = Modifier.padding(getDefaultPadding())
             ) {
-                Snackbar(shape = RoundedCornerShape(50)) {
+                Snackbar(
+                    shape = RoundedCornerShape(50),
+                    action = {
+                        it.visuals.actionLabel?.let { label ->
+                            TextButton(onClick = goToRoot) {
+                                Text(text = label)
+                            }
+                        }
+                    }
+                ) {
                     Text(text = it.visuals.message)
                 }
             }
@@ -132,7 +142,8 @@ fun BackupRestoreSettings(goBack: () -> Unit) {
                             context.resources.getQuantityString(
                                 R.plurals.settings_snackbar_restore,
                                 numberOfRestored, numberOfRestored
-                            )
+                            ),
+                            actionLabel = context.resources.getString(R.string.button_show),
                         )
                     }
                 })
@@ -189,8 +200,8 @@ fun BackupDialog(dismiss: () -> Unit, afterBackup: (numberOfBackups: Int) -> Uni
     val recipesToBackup = remember { mutableStateListOf<Recipe>() }
     val context = LocalContext.current
     val db = AppDatabase.getInstance(context)
-    val recipes by db.recipeDao().getAll().observeAsState(initial = listOf())
-    LaunchedEffect(recipes) {
+    val recipes by db.recipeDao().getAll().observeAsState(initial = emptyList())
+    LaunchedEffect(recipes.isEmpty()) {
         if (recipesToBackup.isEmpty()) {
             recipesToBackup.addAll(recipes)
         }
@@ -215,6 +226,7 @@ fun BackupDialog(dismiss: () -> Unit, afterBackup: (numberOfBackups: Int) -> Uni
         dismiss()
     }
 
+    // TODO: Underlying Dialog doesn't change size when content changes size, maybe rewrite it?
     Material3Dialog(modifier = Modifier.fillMaxSize(), onDismissRequest = dismiss, onSave = {
         val c = Calendar.getInstance().time
         val format: DateFormat =
@@ -230,7 +242,8 @@ fun BackupDialog(dismiss: () -> Unit, afterBackup: (numberOfBackups: Int) -> Uni
                 }
                 ListItem(
                     text = { Text(it.name) },
-                    modifier = Modifier.selectable(selected = isSelected, onClick = onCheck),
+                    modifier = Modifier
+                        .selectable(selected = isSelected, onClick = onCheck),
                     icon = {
                         Checkbox(checked = isSelected, onCheckedChange = { onCheck() })
                     }
@@ -250,17 +263,13 @@ fun DefaultRecipesDialog(dismiss: () -> Unit) {
     val steps = prepopulateData.steps.groupBy { it.recipeId }
     val db = AppDatabase.getInstance(context)
     Material3Dialog(onDismissRequest = dismiss, onSave = {
-        coroutineScope.launch {
-            recipesToAdd.forEach { recipe ->
+        recipesToAdd.forEach { recipe ->
+            coroutineScope.launch {
                 val idOfRecipe = db.recipeDao().insertRecipe(recipe.copy(id = 0))
-                val stepsOfTheRecipe =
-                    steps[prepopulateData.recipes.indexOf(recipe)] ?: return@launch
+                val stepsOfTheRecipe = steps[recipe.id] ?: return@launch
                 db.stepDao().insertAll(
                     stepsOfTheRecipe.map {
-                        it.copy(
-                            id = 0,
-                            recipeId = idOfRecipe.toInt()
-                        )
+                        it.copy(id = 0, recipeId = idOfRecipe.toInt())
                     }
                 )
             }
@@ -290,5 +299,5 @@ fun DefaultRecipesDialog(dismiss: () -> Unit) {
 @Preview
 @Composable
 fun BackupRestoreSettings() {
-    BackupRestoreSettings(goBack = { })
+    BackupRestoreSettings(goBack = {}, goToRoot = {})
 }