+++ docs/Changelog.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - show recipe information on main recipe list screen
 - tweaked layout of timer
 - WYSIWYG in description input
+- Added "Show" button to go to recipe list after restoring backup file
+- Added "Cancel" button to Dialogs that we missing it
 
 ### Changed
 - Back button now closes icon picker bottom sheet in recipe edit
@@ -19,6 +21,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Description input is hidden by default to indicate that it isn't required
 - Changed padding in StepAddCard and Description components
 - Tweaked layout of recipe edit page
+- Fixed adding default recipes
+- Changed settings items position
 
 ### Removed